remove old Nexus publishing.
