+++ src/main/groovy/se/alipsa/gradle/plugin/release/NexusReleasePluginExtension.groovy
@@ -16,7 +16,6 @@ class NexusReleasePluginExtension {
   final Property<String> userName
   final Property<String> password
   final Property<MavenPublication> mavenPublication
-  final Property<PublishingType> publishingType
 
   // ✅ Optional references to tasks
   TaskProvider<Task> bundleTask
@@ -27,14 +26,7 @@ class NexusReleasePluginExtension {
     userName = objects.property(String)
     password = objects.property(String)
     mavenPublication = objects.property(MavenPublication)
-    publishingType = objects.property(PublishingType).convention(PublishingType.CENTRAL)
-    String defaultUrl
-    if (publishingType.get() == PublishingType.CENTRAL) {
-      defaultUrl = CentralPortalClient.CENTRAL_PORTAL_URL
-    } else {
-      defaultUrl = null
-    }
-    nexusUrl = objects.property(String).convention(defaultUrl)
+    nexusUrl = objects.property(String).convention(CentralPortalClient.CENTRAL_PORTAL_URL)
   }
 
   void setNexusUrl(String url) {
@@ -51,10 +43,6 @@ class NexusReleasePluginExtension {
 
   void setMavenPublication(MavenPublication publication) {
     mavenPublication.set(publication)
-  }
-
-  void setPublishingType(PublishingType type) {
-    publishingType.set(type)
   }
 
 }
