--- src/main/groovy/se/alipsa/gradle/plugin/release/PublishingType.groovy
@@ -1,8 +1,0 @@
-package se.alipsa.gradle.plugin.release
-
-import groovy.transform.CompileStatic
-
-@CompileStatic
-enum PublishingType {
-  NEXUS, CENTRAL
-}
